backend/requirements.txt
amqp==5.1.1
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.2
eventlet==0.33.3
Flask==2.3.3
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
fonttools==4.42.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.3.1
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
redis==5.0.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7